+++ src/main.js
@@ -33,7 +33,7 @@ function createWindow() {
     height: 900,
     minWidth: 900,
     minHeight: 600,
-    icon: path.join(__dirname, '..', 'agency.ico'),
+    icon: path.join(__dirname, '..', 'groundcontrol.ico'),
     backgroundColor: bg,
     frame: false,
     titleBarStyle: 'hidden',